Add files via upload
Introduction: This project offers an in-depth exploration of global terrorism events, emphasizing patterns, hotspots, and trends from 1970 to 2017. Terrorism has been an influential global actor, shaping political decisions, socio-economic trends, and international relations. This EDA aims to provide clarity on the nuances of this critical issue.

Objective: To identify the countries most affected by terrorism. To discover patterns concerning the frequency and impact of attacks. To determine the evolution of terrorist activities over the years.